aws/email/requirements.txt
boto3>=1.26.0
dnspython>=2.3.0
